# internal/adapters/inbound/http/openapi.yaml
openapi: 3.1.0
info:
  title: ObservAI API
  version: 1.0.0
  description: |
    Provider-agnostic observability analysis gateway.

    All successful responses are wrapped in `WrapperDtoResponde` with a stable
    `data` and `metadata` envelope. Provider-specific fields never leak into
    the public contract — adapters normalize them into the schemas documented
    below.

    The interaction chat answers only questions related to an active analysis.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /v1/setup/admin:
    post:
      tags: [setup]
      summary: Bootstrap the initial admin user. Rejected once any user exists.
      operationId: bootstrapAdmin
      responses:
        "201": { description: Initial admin created. }
        "409": { description: Setup already completed. }
...